More user-friendly
It is now possible to customize the default :key and :test args.

Also, when a :key argument is given to #'(setf alref), it will
issue a continuable error when a new pair is about to be created.
Previously, the pair would be created silently, introducing potential
bugs to a program.

Future versions of ALREF will probably include support for some
amount of 're-structuring' in this case.